backend/requirements.txt
fastapi==0.111.0
uvicorn[standard]==0.30.1
sqlalchemy==2.0.30
pydantic==2.7.1
pydantic-settings==2.2.1
python-dotenv==1.0.1
anthropic==0.28.0
httpx<0.28.0
requests==2.32.3
python-multipart==0.0.9
aiofiles==23.2.1
slowapi==0.1.9
